backend/requirements.txt
Flask
Flask-Cors
mysql-connector-python
bcrypt
PyJWT
